pyproject.toml
[project]
name = "klin-university"
version = "0.1.0"
description = ""
authors = [
    {name = "Андрей Цуриков",email = "fair20002@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-filter (>=25.2,<26.0)",
    "djoser (>=2.3.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.flake8]
exclude = [
    "**/tests/",
    "src/*/migrations/",
    "manage.py",
    "*/__pycache__/",
    ".venv",
    "venv",
]
per-file-ignores = """
    settings.py:E501
"""


[tool.isort]
profile = "black"
line_length = 79
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
known_first_party = ["employees", "core", "api"]
src_paths = ["src"]
[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-django = "^4.11.1"
factory-boy = "^3.3.3"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "employees.settings"
pythonpath = ["src", "."]
testpaths = ["tests"]